requirements.txt
# Core Web Framework
Flask
flask_migrate
gunicorn
#pymongo[srv]

# IBM watsonx.ai SDK
ibm-watsonx-ai

# For secure API calls and JSON parsing
requests
python-dotenv
pytesseract
Pillow